refactor: remove input port directory for user
