+++ internal/adapters/primary/http/httpserver/builder.go
@@ -6,7 +6,6 @@ import (
 	"github.com/lechitz/AionApi/internal/adapters/primary/http/middleware/auth"
 	"github.com/lechitz/AionApi/internal/adapters/primary/http/middleware/contextbuilder"
 	"github.com/lechitz/AionApi/internal/core/ports/input"
-	"github.com/lechitz/AionApi/internal/core/ports/input/http"
 	"github.com/lechitz/AionApi/internal/core/ports/output/cache"
 	"github.com/lechitz/AionApi/internal/core/ports/output/logger"
 	portRouter "github.com/lechitz/AionApi/internal/core/ports/output/router"
@@ -34,7 +33,7 @@ type RouterBuilder struct {
 //   - error: any error encountered during setup
 func BuildRouterRoutes(
 	logger logger.Logger,
-	userService http.UserService,
+	userService input.UserService,
 	authService input.AuthService,
 	tokenRepo cache.TokenRepositoryPort,
 	contextPath string,
