+++ internal/core/ports/input/user_input.go
@@ -1,4 +1,4 @@
-package http
+package input
 
 import (
 	"context"
